acme: implement ListCertAlternates
Let's Encrypt is defaulting to a longer cross-signed chain on
May 4th, 2021 but will offer the ability to download the shorter chain via
an alternate URL via a link header [1]. The alternate relation is described
in section 7.4.2 of RFC 8555.

ListCertAlternates should be passed the original certificate chain URL and
will return a list of alternate chain URLs that can be passed to FetchCert
to download.

Fixes golang/go#42437

[1] https://community.letsencrypt.org/t/production-chain-changes/150739

Change-Id: Iaa32e49cb1322ac79ac1a5b4b7980d5401f4b86e